Moves some compatibility hacks to johnny.compat
